refactor (core): split LayeredMaterial constants and CommandCancelledEx to their own file
As LayeredMaterial and Scheduler depend on a lot of other module (e.g Capabilities which in
turn depends on 'document') it's desirable to avoid importing these 2 modules.

This commit moves their constant/subclass declarations to their own file to ease unit-testing.
