+++ src/Process/FeatureProcessing.js
@@ -1,6 +1,6 @@
 import * as THREE from 'three';
 import LayerUpdateState from '../Core/Layer/LayerUpdateState';
-import { CancelledCommandException } from '../Core/Scheduler/Scheduler';
+import CancelledCommandException from '../Core/Scheduler/CancelledCommandException';
 import ObjectRemovalHelper from './ObjectRemovalHelper';
 
 function applyOffset(obj, offset) {
