+++ src/Renderer/LayeredMaterialConstants.js
@@ -1,0 +1,3 @@
+export const l_ELEVATION = 0;
+export const l_COLOR = 1;
+export const EMPTY_TEXTURE_ZOOM = -1;
